backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx[http2]==0.26.0

# LiteLLM for multi-provider LLM support
litellm==1.34.0

# MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# Environment management
python-dotenv==1.0.0

# Utilities
cachetools==5.3.2

# Google Generative AI (optional, for direct Gemini access)
google-generativeai==0.3.2
